+++ Sources/GraphQL/Subscription/Subscribe.swift
@@ -1,5 +1,3 @@
-import Dispatch
-import Runtime
 import NIO
 
 /**
